Basic bot for Toka
I started to rebuild BobBot, however I have decided to support Toka.io for this one, and Twitch later on. This should be all the basics that I can get at this point.

It contains a full command hierarchy with mods, universal mods(Highest status), and normal people. Read the comments if you want to, but it's only made for the basics.